release test & fix bug
