+++ .gitignore
@@ -3,3 +3,4 @@ demo2/
 node_modules/
 dist/
 yarn-error.log
+debug.js
