+++ package.json
@@ -2,11 +2,11 @@
   "name": "md-converter",
   "version": "0.0.3",
   "description": "Md Converter make markdown convert to html",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "build": "webpack",
     "debug": "yarn build; node ./dist/index.js",
-    "test": "jest"
+    "test": "jest --config=jest.config.js"
   },
   "author": "nappa <masa.yanagawa721@gmail.com>",
   "repository": {
@@ -24,7 +24,9 @@
     "babel-jest": "^26.6.3",
     "babel-loader": "^8.2.2",
     "clean-webpack-plugin": "^3.0.0",
+    "css-loader": "^5.0.1",
     "jest": "^26.6.3",
+    "mini-css-extract-plugin": "^1.3.3",
     "webpack": "^5.11.0",
     "webpack-cli": "^4.2.0"
   },
