+++ webpack.config.js
@@ -1,10 +1,11 @@
 const path = require('path');
 const { CleanWebpackPlugin } = require('clean-webpack-plugin');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 
 const ROOT_DIR = __dirname;
 const SRC_DIR = path.resolve(ROOT_DIR, 'src');
 const DIST_DIR = path.resolve(ROOT_DIR, 'dist');
-const ENTRY_FILE = path.resolve(SRC_DIR, 'main.js');
+const ENTRY_FILE = path.resolve(SRC_DIR, 'index.js');
 const OUT_FILENAME = 'index.js'
 
 module.exports = {
@@ -21,15 +22,17 @@ module.exports = {
         exclude: /node_modules/,
         use: [ 'babel-loader' ]
       },
+      {
+        test: /\.css$/,
+        use: [ MiniCssExtractPlugin.loader, 'css-loader' ]
+      }
     ]
   },
   resolve: {
     extensions: [ '.js' ],
-    alias: {
-      '@': SRC_DIR
-    }
   },
   plugins: [
     new CleanWebpackPlugin(),
+    new MiniCssExtractPlugin(),
   ],
 }
